Rename DeviceStoryDevice
