+++ app/controllers/device_stories_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class DeviceStoriesController < ApplicationController
+  has_scope :order
+
+  def index
+    @device_stories = apply_scopes(DeviceStory)
+      .page(params[:page])
+      .per(params[:per_page])
+  end
+end
