--- app/controllers/device_story/devices_controller.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module DeviceStory
-  class DevicesController < ApplicationController
-    has_scope :order
-
-    def index
-      @devices = apply_scopes(::DeviceStory::Device)
-        .page(params[:page])
-        .per(params[:per_page])
-    end
-  end
-end
