--- app/views/device_story/devices/index.html.erb
@@ -1,34 +1,0 @@
-<%= content_for(:title) { "Device Story: Devices" } %>
-
-<h2>Device Story: Devices</h2>
-
-<div class="well">
-  <%= number_with_delimiter @devices.total_count %> <%= t('.results') %>.
-</div>
-
-<%- if @devices.empty? -%>
-  <p><%= t(".none_found") -%></p>
-<%- else -%>
-  <table class="table table-hover" data-provides="rowlink">
-    <thead>
-      <%= table_sort_header(:device, :device_urn) %>
-      <%= table_sort_header(:device, :last_values) %>
-      <th>Location</th>
-      <th>Last seen</th>
-    </thead>
-    <tbody>
-      <%- @devices.each do |device| -%>
-        <tr>
-          <td><%= device.device_urn %></td>
-          <td><%= device.last_values %></td>
-          <td><%= device.location %></td>
-          <td><%= device.last_seen_formatted %></td>
-        </tr>
-      <%- end -%>
-    </tbody>
-  </table>
-  <%= paginate @devices, theme: 'twitter-bootstrap-2' -%>
-<%- end -%>
-
-<div class="spacer"></div>
-<%= current_page_api_example :device_story_devices_url %>
